Rename uWSGI environment variables
According to
https://uwsgi-docs.readthedocs.io/en/latest/Configuration.html#environment-variables,
uWSGI takes any environment variable that starts with "UWSGI_" and
assigns a configuration directive based on the name of the variable.
This causes errors when you have strict mode turned on with `strict =
true' if you have a variable (say "UWSGI_THREAD_COUNT") that's set for
whatever reason, perhaps to render a config file template, that does not
translate to a valid configuration directive.

To avoid this problem, we will rename the variables to start with
"U_WSGI" so that they are not interpreted as directives by uWSGI.